pyproject.toml
[tool.poetry]
name = "truss-calculator"
version = "0.1.0"
description = "A simple truss calculator system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"